% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.data.frame.patch}
\alias{as.data.frame.patch}
\title{Convert a patch object to a data frame}
\usage{
\method{as.data.frame}{patch}(x, row.names = NULL, ..., long = TRUE)
}
\arguments{
\item{x}{object of class `patch`}

\item{row.names}{name of patch. Note that this will be a column in
the returned data.frame. If NULL, name is set to 1. If a vector of
length greater than 1 is given, only the first name is used.}

\item{...}{any other arguments. not used}

\item{long}{whether the data.frame should be long (1 row for each
variable in the object of class `patch`), or wide (a single row
where each variable in the object of class patch `patch` is a
column.)}
}
\value{
data.frame
}
\description{
..
}
\author{
Sangeeta Bhatia
}
