% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.data.frame.state}
\alias{as.data.frame.state}
\title{Conversion of state object}
\usage{
\method{as.data.frame}{state}(x, row.names = NULL, ..., long = TRUE)
}
\arguments{
\item{x}{object of class `state`}

\item{row.names}{names to be assigned to patches. If NULL,
names will be numeric from 1 to number of patches. Note that these
will actually be column values in the data.frame, rather than
row.names.}

\item{...}{unused arguments}

\item{long}{if true, data frame is in long format with the columns
patch, variable, and value. Column patch contains the patch name
assigned through row.names (1:number of patches, if row.names
is NULL. the column variable are the elements of patch (
susceptible, exposed, infected, recovred, transmission_rate,
recovery_rate, birth_rate, death_rate
))}
}
\value{
data.frame
}
\description{
Convert state object to a data.frame
}
\author{
Sangeeta Bhatia
}
