% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{is.patch}
\alias{is.patch}
\title{Test for class patch}
\usage{
is.patch(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{optional arguments not used}
}
\value{
TRUE if x is of class patch, FALSE otherwise
}
\description{
Test for class patch
}
\author{
Sangeeta Bhatia
}
