% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{is.state}
\alias{is.state}
\title{Test for class state}
\usage{
is.state(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{optional arguments not used}
}
\value{
TRUE of x if of class state, FALSE
otherwise.
}
\description{
Test for class state
}
\author{
Sangeeta Bhatia
}
