% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{make_patch}
\alias{make_patch}
\title{Constructor for patch class}
\usage{
make_patch(s_patch, e_patch, i_patch, r_patch, birth_rate, death_rate,
  transmission_rate, infection_rate, recovery_rate)
}
\arguments{
\item{s_patch}{Number of susceptible individuals in patch}

\item{e_patch}{Number of exposed individuals in patch}

\item{i_patch}{Number of infected individuals in patch}

\item{r_patch}{Number of recovered individuals in patch}

\item{birth_rate}{Birth rate for this patch. Number of births per unit time}

\item{death_rate}{Death rate for this patch. Number of deaths per unit time in this patch}
}
\value{
a list of class patch which has the following items
** susceptible ** Number of susceptibles in this patch
** exposed ** Number of exposed individuals in this patch
** infected ** Number of infected individuals in this patch
** recovered ** Number of recovered individuals in this patch
** birth_rate ** Patch-specific birth rate
** death_rate ** Patch-specific death rate
}
\description{
Build a patch with seir compartments and patch-specific birth and death rates
}
\author{
Sangeeta Bhatia
}
