% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch}
\alias{patch}
\title{Constructor for patch class}
\usage{
patch(s_patch, e_patch, i_patch, r_patch, birth_rate, death_rate,
  transmission_rate, infection_rate, recovery_rate)
}
\arguments{
\item{s_patch}{Number of susceptible individuals in patch}

\item{e_patch}{Number of exposed individuals in patch}

\item{i_patch}{Number of infected individuals in patch}

\item{r_patch}{Number of recovered individuals in patch}

\item{birth_rate}{Birth rate for this patch. Number of births per unit time}

\item{death_rate}{Death rate for this patch. Number of deaths per unit time in this patch}

\item{transmission_rate}{effective contact rate}

\item{infection_rate}{rate of moving from compartment E to I}

\item{recovery_rate}{rate of moving from compartment I to R.}
}
\value{
a list of class patch which has the following items
** susceptible ** Number of susceptibles in this patch
** exposed ** Number of exposed individuals in this patch
** infected ** Number of infected individuals in this patch
** recovered ** Number of recovered individuals in this patch
** birth_rate ** Patch-specific birth rate
** death_rate ** Patch-specific death rate
** transmission_rate ** Effective contact rate for this patch
** Infection rate ** Rate at which individuals move from E to I
** Recovery rate ** Rate at which individuals move from  I to R.
}
\description{
Build a patch with seir compartments and patch-specific birth and death rates
}
\author{
Sangeeta Bhatia
}
