% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_symptoms}
\alias{patch_symptoms}
\title{Constructor for patch class}
\usage{
patch_symptoms(s_patch, e_patch, i_a_patch, i_p_patch, i_s_patch, r_patch,
  birth_rate, death_rate, transmission_rate, asymptomatic_infectiousness,
  presymptomatic_infectiousness, prop_symptomatic, infection_rate,
  symptom_rate, recovery_rate_asym, recovery_rate_sym)
}
\arguments{
\item{s_patch}{Number of susceptible individuals in patch}

\item{e_patch}{Number of exposed individuals in patch}

\item{i_a_patch}{Number of asymptomatic infected individuals in patch}

\item{i_p_patch}{Number of presymptomatic infected individuals in patch}

\item{i_s_patch}{Number of symptomatic infected individuals in patch}

\item{r_patch}{Number of recovered individuals in patch}

\item{birth_rate}{Birth rate for this patch. Number of births per unit time}

\item{death_rate}{Death rate for this patch. Number of deaths per unit time in this patch}

\item{transmission_rate}{effective contact rate}

\item{asymptomatic_infectiousness}{Proportion giving 
the infectivity of asymptomatic cases relative to
symptomatic cases.}

\item{presymptomatic_infectiousness}{Proportion giving 
the infectivity of presymptomatic cases relative to
symptomatic cases.}

\item{prop_symptomatic}{Proportion of cases that are symptomatic.}

\item{infection_rate}{rate of moving from compartment E to I}

\item{symptom_rate}{Rate of moving from I_presymptomatic to I_symptomatic.}

\item{recovery_rate_asym}{Rate of moving from I_asymtpomatic to recovered}

\item{recovery_rate_sym}{Rate of moving from I_symtpomatic to recovered}
}
\value{
a list of class patch which has the following items
** susceptible ** Number of susceptibles in this patch
** exposed ** Number of exposed individuals in this patch
** infected_asymptomatic ** Number of asymptomatic infected individuals in this patch
** infected_presymptomatic ** Number of presymptomatic infected individuals in this patch
** infected_symptomatic ** Number of symptomatic infected individuals in this patch
** recovered ** Number of recovered individuals in this patch
** birth_rate ** Patch-specific birth rate
** death_rate ** Patch-specific death rate
** transmission_rate ** Effective contact rate for this patch
** asymptomatic_infectiousness ** Relative infectivity of asymptomatic infected individuals
** presymptomatic_infectiousness ** Relative infectivity of presymptomatic infected individuals
** prop_symptomatic ** Proportion of infections with symptoms
** Infection rate ** Rate at which individuals move from E to I_a or I_p
** Symptom rate ** Rate at which individuals move from I_p to I_s
** Asymptomatic recovery rate ** Rate at which individuals move from  I_a to R.
** Symptomatic recovery rate ** Rate at which individuals move from  I_s to R.
}
\description{
Build a patch with seir compartments, including (a/pre-)symptomatic compartments,
and patch-specific birth and death rates
}
\author{
Jack Wardle
}
