% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_symptoms_testing}
\alias{patch_symptoms_testing}
\title{Constructor for patch class}
\usage{
patch_symptoms_testing(s_patch, e_patch, i_a_patch, i_p_patch, i_s_patch,
  r_patch, e_diag_patch, i_a_diag_patch, i_p_diag_patch, i_s_diag_patch,
  r_diag_patch, s_false_patch, r_false_patch, birth_rate, death_rate,
  transmission_rate, transmission_rate_pilgrims,
  transmission_rate_pilgrims_and_atrisk, asymptomatic_infectiousness,
  presymptomatic_infectiousness, prop_symptomatic, infection_rate,
  symptom_rate, recovery_rate_asym, recovery_rate_sym, testing_rate,
  test_sensitivity, test_specificity, isolation_period)
}
\arguments{
\item{s_patch}{Number of susceptible individuals in patch}

\item{e_patch}{Number of exposed individuals in patch}

\item{i_a_patch}{Number of asymptomatic infected individuals in patch}

\item{i_p_patch}{Number of presymptomatic infected individuals in patch}

\item{i_s_patch}{Number of symptomatic infected individuals in patch}

\item{r_patch}{Number of recovered individuals in patch}

\item{e_diag_patch}{Number of diagnosed exposed individuals in patch}

\item{i_a_diag_patch}{Number of diagnosed asymptomatic infected individuals in patch}

\item{i_p_diag_patch}{Number of diagnosed presymptomatic infected individuals in patch}

\item{i_s_diag_patch}{Number of diagnosed symptomatic infected individuals in patch}

\item{r_diag_patch}{Number of diagnosed recovered individuals in patch}

\item{s_false_patch}{Number of susceptible individuals with false positive test in patch}

\item{r_false_patch}{Number of recovered individuals with false positive test in patch}

\item{birth_rate}{Birth rate for this patch. Number of births per unit time}

\item{death_rate}{Death rate for this patch. Number of deaths per unit time in this patch}

\item{transmission_rate}{effective contact rate}

\item{transmission_rate_pilgrims}{effective contact rate between pilgrims}

\item{transmission_rate_pilgrims_and_atrisk}{effective contact rate between pilgrims and at risk non-pilgrims}

\item{asymptomatic_infectiousness}{Proportion giving 
the infectivity of asymptomatic cases relative to
symptomatic cases.}

\item{presymptomatic_infectiousness}{Proportion giving 
the infectivity of presymptomatic cases relative to
symptomatic cases.}

\item{prop_symptomatic}{Proportion of cases that are symptomatic.}

\item{infection_rate}{rate of moving from compartment E to I}

\item{symptom_rate}{Rate of moving from I_presymptomatic to I_symptomatic.}

\item{recovery_rate_asym}{Rate of moving from I_asymptomatic to recovered}

\item{recovery_rate_sym}{Rate of moving from I_symptomatic to recovered}

\item{testing_rate}{Proportion of people that are tested}

\item{test_sensitivity}{Sensitivity of tests}

\item{test_specificity}{Specificity of tests}

\item{isolation_period}{Average duration of isolation following positive test}
}
\value{
a list of class patch which has the following items
** susceptible ** Number of susceptibles in this patch
** exposed ** Number of exposed individuals in this patch
** exposed_diagnosed ** Number of diagnosed exposed individuals in this patch
** infected_asymptomatic ** Number of asymptomatic infected individuals in this patch
** infected_asymptomatic_diagnosed ** Number of diagnosed asymptomatic infected individuals in this patch
** infected_presymptomatic ** Number of presymptomatic infected individuals in this patch
** infected_presymptomatic_diagnosed ** Number of diagnosed presymptomatic infected individuals in this patch
** infected_symptomatic ** Number of symptomatic infected individuals in this patch
** infected_symptomatic_diagnosed ** Number of diagnosed symptomatic infected individuals in this patch
** recovered ** Number of recovered individuals in this patch
** recovered_diagnosed ** Number of diagnosed recovered individuals in this patch
** birth_rate ** Patch-specific birth rate
** death_rate ** Patch-specific death rate
** transmission_rate ** Effective contact rate between non-pilgrims
** transmission_rate_pilgrims ** Effective contact rate between pilgrims
** transmission_rate_pilgrims_and_atrisk ** Effective contact rate between pilgrims and "at risk" non-pilgrims
** asymptomatic_infectiousness ** Relative infectivity of asymptomatic infected individuals
** presymptomatic_infectiousness ** Relative infectivity of presymptomatic infected individuals
** prop_symptomatic ** Proportion of infections with symptoms
** Infection rate ** Rate at which individuals move from E to I_a or I_p
** Symptom rate ** Rate at which individuals move from I_p to I_s
** Asymptomatic recovery rate ** Rate at which individuals move from  I_a to R.
** Symptomatic recovery rate ** Rate at which individuals move from  I_s to R.
** Testing rate ** Proportion of individuals that are tested.
** Test sensitivity ** Sensitivity of tests.
** Test specificity ** Specificity of tests.
** Isolation period ** Average duration of isolation period in days.
}
\description{
Build a patch with seir compartments, including (a/pre-)symptomatic compartments,
compartments for diagnosed individuals, and patch-specific birth and death rates
}
\author{
Jack Wardle
}
