% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_state_supporting_functions.R
\name{simulate_isolation_period}
\alias{simulate_isolation_period}
\title{Simulate Isolation Period}
\usage{
simulate_isolation_period(starting_compartment, input_state, sample_size,
  isolation_period)
}
\arguments{
\item{starting_compartment}{The starting compartment.}

\item{input_state}{The input state generated using state_symptoms_testing().}

\item{sample_size}{The sample size to start with.}

\item{isolation_period}{The isolation period.}
}
\value{
A vector of probabilities for each final compartment.
}
\description{
Simulate Isolation Period
}
