% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state}
\alias{state}
\title{Constructor for state class}
\usage{
state(s_patches, e_patches, i_patches, r_patches, birth_rates, death_rates,
  transmission_rates, infection_rates, recovery_rates, movement_rate)
}
\arguments{
\item{s_patches}{Vector of number of
susceptibles in each patch,}

\item{e_patches}{Vector of number of
exposed in each patch}

\item{i_patches}{Vector of number of
infected in each patch}

\item{r_patches}{Vector of number of
recovered in each patch}

\item{birth_rates}{patch-specific birth rates}

\item{death_rates}{patch-specific death rates}

\item{transmission_rates}{Vector of patch-specific transmission
rate}

\item{infection_rates}{Vector of patch-specific infection rate}

\item{recovery_rates}{Vector of patch-specific recovery rate}

\item{movement_rate}{M rate of movement between
patches is defined the rate matrix M where
m[i, j] is the rate of movement from i to j
conditional on moving out of i per unit time.
Must be n X n matrix where is the number of
patches i.e, max of lengths of all input
vectors s_patches, e_patches, i_patches,
r_patches, birth_rates, and death_rates.}
}
\value{
List with class `state`. Each element
of `state` is a list of class `patch`.
In addition, `state` contains a `movement_rate`
which is a non-negative matrix of rates of
movement between patches.
}
\description{
Build a state, which is a collection of
patches where each patch has SEIR
compartments and patch-specific birth and
death rates.
}
\seealso{
[patch()]
}
\author{
Sangeeta Bhatia
}
