% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_symptoms}
\alias{state_symptoms}
\title{Constructor for state class}
\usage{
state_symptoms(s_patches, e_patches, i_a_patches, i_p_patches, i_s_patches,
  r_patches, birth_rates, death_rates, transmission_rates,
  asymptomatic_infectiousness, presymptomatic_infectiousness, prop_symptomatic,
  infection_rates, symptom_rates, recovery_rates_asym, recovery_rates_sym,
  movement_rate)
}
\arguments{
\item{s_patches}{Vector of number of
susceptibles in each patch,}

\item{e_patches}{Vector of number of
exposed in each patch}

\item{i_a_patches}{Vector of number of
asymptomatic infected in each patch}

\item{i_p_patches}{Vector of number of
pre-symptomatic infected in each patch}

\item{i_s_patches}{Vector of number of
symptomatic infected in each patch}

\item{r_patches}{Vector of number of
recovered in each patch}

\item{birth_rates}{patch-specific birth rates}

\item{death_rates}{patch-specific death rates}

\item{transmission_rates}{Vector of patch-specific transmission
rate}

\item{asymptomatic_infectiousness}{Vector of proportions that 
give the infectivity of asymptomatic cases relative to
symptomatic cases.}

\item{presymptomatic_infectiousness}{Vector of proportions that 
give the infectivity of presymptomatic cases relative to
symptomatic cases.}

\item{prop_symptomatic}{Vector of patch-specific proportion
of cases that are symptomatic}

\item{infection_rates}{Vector of patch-specific infection rate}

\item{symptom_rates}{Vector of patch-specific rate at which
pre-symptomatic infections develop symptoms}

\item{recovery_rates_asym}{Vector of patch-specific recovery rate
for asymptomatic infections}

\item{recovery_rates_sym}{Vector of patch-specific recovery rate
for symptomatic infections}

\item{movement_rate}{M rate of movement between
patches is defined the rate matrix M where
m[i, j] is the rate of movement from i to j
conditional on moving out of i per unit time.
Must be n X n matrix where is the number of
patches i.e, max of lengths of all input
vectors s_patches, e_patches, i_patches,
r_patches, birth_rates, and death_rates.}
}
\value{
List with class `state`. Each element
of `state` is a list of class `patch`.
In addition, `state` contains a `movement_rate`
which is a non-negative matrix of rates of
movement between patches.
}
\description{
Build a state, which is a collection of
patches where each patch has SEIR
compartments, with the addition of asymptomatic,
pre-symptomatic and symptomatic compartments,
and patch-specific birth and
death rates.
}
\seealso{
[patch()]
}
\author{
Jack Wardle
}
