% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_patch.R
\name{update_state}
\alias{update_state}
\title{Update state}
\usage{
update_state(state, dt, compartments = c("susceptible", "exposed",
  "infected", "recovered"), movement_type = c("rate", "probability"))
}
\arguments{
\item{state}{state is a collection of patches and a matrix of
rates of movement between patches.}

\item{dt}{time for which state should be updated. It is the user's
responsibility to make sure that this number is consistent
with the units on rates. For instance, if the various rates are
per week, dt is assumed to be dt weeks.}

\item{compartments}{in case they are different from SEIR}

\item{movement_type}{select whether the movement matrix input
contains rates (with the caveat on appropriate units from
above remaining valid) or probabilities.}
}
\value{
state updated
}
\description{
Update state
}
\author{
Sangeeta Bhatia
}
